DRY things up pt I
